requirements.txt
python-telegram-bot==20.8
telethon==1.34.0
aiohttp==3.9.5
